+++ src/events.rs
@@ -1,24 +1,29 @@
 use std::pin::pin;
 use std::time::Duration;
 
+use color_eyre::Result;
 use crossterm::event::KeyModifiers;
 use crossterm::event::{Event as CrossEvent, EventStream, KeyCode, KeyEventKind};
 use enum_as_inner::EnumAsInner;
 use futures::{Stream, StreamExt, stream};
-use paho_mqtt::{AsyncClient, QoS};
+use rumqttc::{AsyncClient, ConnectionError, EventLoop, Incoming, QoS};
 use tokio::sync::mpsc::UnboundedReceiver;
 use tokio::sync::mpsc::unbounded_channel;
 use tokio::task;
 use tokio::time::sleep;
 
-use crate::model::Message;
-
 const TICK: Duration = Duration::from_millis(100);
 
-#[derive(Debug, PartialEq, EnumAsInner)]
+#[derive(Debug, EnumAsInner)]
 pub enum Event {
     Render(RenderEvent),
     Update(UpdateEvent),
+}
+impl Event {
+    pub fn is_disconnect(&self) -> bool {
+        self.as_render()
+            .is_some_and(|e| e == &RenderEvent::Disconnect)
+    }
 }
 
 #[derive(Debug, PartialEq, Eq)]
@@ -36,26 +41,54 @@ pub enum RenderEvent {
     End,
     Select,
     Quit,
+    Connect,
+    Disconnect,
 }
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug)]
 pub enum UpdateEvent {
-    Receive(Message),
+    Receive(rumqttc::Publish),
 }
 
-pub async fn handler(client: &mut AsyncClient) -> UnboundedReceiver<Event> {
+pub async fn start(
+    client: AsyncClient,
+    mut eventloop: EventLoop,
+) -> Result<UnboundedReceiver<Event>> {
     let (tx, rx) = unbounded_channel();
-    client.subscribe("#", QoS::default());
-
-    let stream = client.get_stream(None);
+    let tx2 = tx.clone();
     task::spawn(async move {
-        let events = stream::select(keys(), messages(stream));
-        let mut events = pin!(stream::select(events, tick()));
+        loop {
+            if eventloop.poll().await.is_err() {
+                sleep(Duration::from_millis(500)).await;
+                let _ = tx2.send(Event::Render(RenderEvent::Disconnect));
+                continue;
+            }
+
+            let _ = client.subscribe("#", QoS::AtMostOnce).await;
+            let _ = tx2.send(Event::Render(RenderEvent::Connect));
+            loop {
+                match eventloop.poll().await {
+                    Err(ConnectionError::Io(_)) => {
+                        let _ = tx2.send(Event::Render(RenderEvent::Disconnect));
+                        break;
+                    }
+                    Ok(rumqttc::Event::Incoming(Incoming::Publish(message))) => {
+                        let _ = tx2.send(Event::Update(UpdateEvent::Receive(message)));
+                    }
+                    _ => {}
+                }
+            }
+        }
+    });
+
+    task::spawn(async move {
+        let mut events = pin!(stream::select(keys(), tick()));
         while let Some(event) = events.next().await {
             let _ = tx.send(event);
         }
     });
-    rx
+
+    Ok(rx)
 }
 
 fn tick() -> impl Stream<Item = Event> {
@@ -63,12 +96,6 @@ fn tick() -> impl Stream<Item = Event> {
         sleep(TICK).await;
         Some((Event::Render(RenderEvent::Tick), ()))
     })
-}
-
-fn messages(stream: impl Stream<Item = Option<paho_mqtt::Message>>) -> impl Stream<Item = Event> {
-    stream
-        .filter_map(|message| async move { message })
-        .map(|message| Event::Update(UpdateEvent::Receive(message.into())))
 }
 
 fn keys() -> impl Stream<Item = Event> {