+++ src/main.rs
@@ -5,14 +5,13 @@ use color_eyre::{
 };
 use mqttui::{
     config::Config,
-    events::{self, Event},
+    events::{self},
     model::Model,
     ui,
 };
-use paho_mqtt::{AsyncClient, ConnectOptions, CreateOptionsBuilder};
 use petname::petname;
 use ratatui::{Terminal, prelude::Backend};
-use tokio::sync::mpsc::UnboundedReceiver;
+use rumqttc::{AsyncClient, EventLoop, MqttOptions};
 use url::Url;
 
 /// Mqtt TUI
@@ -24,33 +23,38 @@ struct Args {
     #[arg(short('b'), long, default_value = "mqtt://localhost:1883", value_parser = validate_url)]
     broker: Url,
 
-    /// Print the path of the default config
+    /// Immediately quit, when the connection to the broker is lost
+    ///
+    /// By default the TUI will stay open and try to automatically reconnect when the broker comes
+    /// back online
+    #[arg(short('q'), long)]
+    quit: bool,
+
+    /// Print the path of where the config file is read from
     #[arg(long)]
-    print_config: bool,
+    print_config_path: bool,
 }
 
 #[tokio::main]
 async fn main() -> Result<()> {
     color_eyre::install()?;
     let args = Args::parse();
-    if args.print_config {
+    if args.print_config_path {
         println!("{}", Config::path()?.display());
         return Ok(());
     }
-    let mut client = init(args.broker).await?;
-    let rx = events::handler(&mut client).await;
 
     let mut terminal = ratatui::init();
     // use ratatui::backend::TestBackend;
     // let mut terminal = Terminal::new(TestBackend::new(10, 10)).unwrap();
 
-    let result = run(&mut terminal, rx).await;
+    let result = run(args.broker, &mut terminal, !args.quit).await;
     ratatui::restore();
     result?;
     Ok(())
 }
 
-async fn init(broker: Url) -> Result<AsyncClient> {
+async fn init(broker: &Url) -> (AsyncClient, EventLoop) {
     let hook = std::panic::take_hook();
     std::panic::set_hook(Box::new(move |panic_info| {
         ratatui::restore();
@@ -58,32 +62,31 @@ async fn init(broker: Url) -> Result<AsyncClient> {
     }));
 
     let name = env!("CARGO_PKG_NAME");
+    let name = format!("{name}-{}", petname(2, "-").unwrap());
 
-    let client = AsyncClient::new(
-        CreateOptionsBuilder::new()
-            .server_uri(broker.clone())
-            .client_id(format!("{name}-{}", petname(2, "-").unwrap()))
-            .finalize(),
-    )?;
-    client
-        .connect(ConnectOptions::default())
-        .await
-        .context(broker)
-        .context("Failed to connect to MQTT broker")?;
-    Ok(client)
+    let mut options = MqttOptions::new(
+        name,
+        broker.host_str().unwrap(),
+        broker.port_or_known_default().unwrap_or(1883),
+    );
+    options.set_max_packet_size(1000000, 1024);
+    let (client, eventloop) = AsyncClient::new(options, 10);
+    (client, eventloop)
 }
 
-async fn run<B: Backend>(
-    terminal: &mut Terminal<B>,
-    mut events: UnboundedReceiver<Event>,
-) -> Result<()> {
-    let mut model = Model::new()?;
+async fn run<B: Backend>(broker: Url, terminal: &mut Terminal<B>, reconnect: bool) -> Result<()> {
+    let (client, eventloop) = init(&broker).await;
+    let mut events = events::start(client, eventloop).await?;
+    let mut model = Model::new(broker.clone())?;
     while !model.shutdown {
         let event = events.recv().await.ok_or(eyre!("runtime died"))?;
         let rendering = event.is_render();
 
+        if !reconnect && event.is_disconnect() {
+            return Err(eyre!(broker)).context("Connection to broker lost");
+        }
+
         model.update(event);
-        // eprintln!("{model:#?}");
 
         if rendering {
             terminal.draw(|frame| ui::render(frame, &model))?;